Move "isType" to Tag.php
